move TLSOption to tls dir
