+++ global/tls_config.go
@@ -48,29 +48,3 @@ func (c *TLSConfig) Clone() *TLSConfig {
 		TLSServerName: c.TLSServerName,
 	}
 }
-
-type TLSOption func(*TLSConfig)
-
-func WithTLS_CACertFile(file string) TLSOption {
-	return func(cfg *TLSConfig) {
-		cfg.CACertFile = file
-	}
-}
-
-func WithTLS_TLSCertFile(file string) TLSOption {
-	return func(cfg *TLSConfig) {
-		cfg.TLSCertFile = file
-	}
-}
-
-func WithTLS_TLSKeyFile(file string) TLSOption {
-	return func(cfg *TLSConfig) {
-		cfg.TLSKeyFile = file
-	}
-}
-
-func WithTLS_TLSServerName(name string) TLSOption {
-	return func(cfg *TLSConfig) {
-		cfg.TLSServerName = name
-	}
-}
